+++ source/extensions/isaacsim.ucx.nodes/docs/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to the Isaac Sim UCX Nodes extension will be documented in this file.
 
+## [1.1.0] - 2025-11-25
+### Added
+- Joint state publishing node
+- Joint command subscribing node
+- Odometry publishing node
+- IMU publishing node
+- RGB image publishing node
+- Camera helper node for connecting the replicator pipeline with image publishing
+
 ## [1.0.0] - 2025-11-18
 
 ### Added